msm: camera: update debugfs entries
Add missing debugfs entries which got removed in
commit 38fe1934c896 'common: Enable debugfs compile check'
and other error checks.

CRs-Fixed: 3280235

+++ drivers/cam_isp/cam_isp_context.h
@@ -1,6 +1,7 @@
 /* SPDX-License-Identifier: GPL-2.0-only */
 /*
  * Copyright (c) 2017-2021, The Linux Foundation. All rights reserved.
+ * Copyright (c) 2022 Qualcomm Innovation Center, Inc. All rights reserved.
  */
 
 #ifndef _CAM_ISP_CONTEXT_H_
@@ -118,7 +119,7 @@ enum cam_isp_state_change_trigger {
 struct cam_isp_ctx_debug {
 	struct dentry  *dentry;
 	uint32_t        enable_state_monitor_dump;
-	uint32_t        enable_cdm_cmd_buff_dump;
+	uint8_t         enable_cdm_cmd_buff_dump;
 };
 
 /**
